added class convert
