+++ data/tasks.txt
@@ -1,4 +1,3 @@
 T|1|read book
-D|0|return book|Monday
-E|0|meeting|U-Town
-P|0|lecture|1600|1800
+D|1|return book|09/09/09 2359
+E|0|meeting|here
