+++ src/main/java/duke/task/Todo.java
@@ -1,6 +1,7 @@
 package duke.task;
 
-import duke.parser.Parser;
+import duke.parser.Convert;
+
 import java.util.Date;
 
 /**
@@ -20,7 +21,7 @@ public class Todo extends Task {
 
     @Override
     public void setNewDate(String date) {
-        this.date = Parser.stringToDate(date);
+        this.date = Convert.stringToDate(date);
     }
 
     @Override
